+++ src/test/java/com/prgrms/shortener/domain/implementation/JpaShortenedUrlFactoryTest.java
@@ -1,8 +1,8 @@
-package com.prgrms.shortener.domain;
+package com.prgrms.shortener.domain.implementation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.prgrms.shortener.domain.implementation.JpaShortenedUrlFactory;
+import com.prgrms.shortener.domain.ShortenedUrl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
