+++ src/test/java/com/prgrms/shortener/domain/implementation/JpaShortenedUrlRepositoryTest.java
@@ -1,8 +1,8 @@
-package com.prgrms.shortener.domain;
+package com.prgrms.shortener.domain.implementation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.prgrms.shortener.domain.implementation.JpaShortenedUrlRepository;
+import com.prgrms.shortener.domain.ShortenedUrl;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
